refactor: Moved classed to appropriate folders
