+++ TrainRoutes/Enums/CountCondition.cs
@@ -1,4 +1,4 @@
-namespace TrainRoutes;
+namespace TrainRoutes.Enums;
 
 public enum CountCondition
 {
