+++ TrainRoutes/FileReader.cs
@@ -7,22 +7,12 @@ public class FileReader
     /// </summary>
     /// <param name="filePath">Input file path location in the project</param>
     /// <returns>The graph structure as provided by the input file</returns>
-    public static Graph ReadGraphFromFile(string filePath)
+    public static string[] ReadGraphFromFile(string filePath)
     {
-        var graph = new Graph();
-        var lines = File.ReadAllLines(filePath);
-
-        foreach (var line in lines)
+        if (File.Exists(filePath))
         {
-            var parts = line.Split(',');
-            var fromTown = parts[0].Trim();
-            var toTown = parts[1].Trim();
-            var distance = int.Parse(parts[2].Trim());
-
-            var route = new Route(fromTown, toTown, distance);
-            graph.AddRoute(route);
+            return File.ReadAllLines(filePath);
         }
-
-        return graph;
+        throw new FileNotFoundException($"File with path {filePath} not found!");
     }
 }
