+++ TrainRoutes/Strategies/TripStrategy/StopBasedTripStrategy.cs
@@ -1,4 +1,6 @@
-namespace TrainRoutes;
+using TrainRoutes.Enums;
+
+namespace TrainRoutes.Strategies.TripStrategy;
 
 public class StopBasedTripStrategy : ITripStrategy
 {
